-changed the container for entities to map, since deletions will happen often +entities can now attack each other and kill each other

+++ GodSimulator/GodSimulator/RandomGenerator.h
@@ -16,7 +16,7 @@ public:
 
 	double GenerateRandomSignedNumber(int nUpperRange);
 
-	double GenerateRandomNumber(int nUpperRange);
+	int GenerateRandomNumber(int nUpperRange);
 
 	int GenerateRandomNumberInRange(int nLowerRange, int nUpperRange);
